Refactor Spell Selection List to new Control
